+++ src/main/regex.c
@@ -4797,7 +4797,7 @@ static bin_tree_t *parse_dup_op(bin_tree_t *elem, re_string_t *regexp, re_dfa_t 
         old_tree = NULL;
 
     if (elem->token.type == SUBEXP)
-        postorder(elem, mark_opt_subexp, (void *)(long)elem->token.opr.idx);
+        postorder(elem, mark_opt_subexp, (void *)(uintptr_t)elem->token.opr.idx);
 
     tree = create_tree(dfa, elem, NULL, (end == -1 ? OP_DUP_ASTERISK : OP_ALT));
     if (BE(tree == NULL, 0))
@@ -5946,7 +5946,7 @@ static bin_tree_t *create_token_tree(re_dfa_t *dfa, bin_tree_t *left, bin_tree_t
 
 static reg_errcode_t mark_opt_subexp(void *extra, bin_tree_t *node)
 {
-    int idx = (int)(long)extra;
+    int idx = (int)(uintptr_t)extra;
     if (node->token.type == SUBEXP && node->token.opr.idx == idx)
         node->token.opt_subexp = 1;
